tutorial-39

+++ database/migrations/2019_08_28_110443_add_category_id_to_posts.php
@@ -14,7 +14,7 @@ class AddCategoryIdToPosts extends Migration
     public function up()
     {
         Schema::table('posts', function (Blueprint $table) {
-            $table->integer('category_id')->nullable()->after('slug')->
+            $table->BigInteger('category_id')->nullable()->after('slug')->
             unsigned(); //unsigned intiger dozvoljava samo dodavanje pozitivnih brojeva
         });
     }
